+++ packages/ui/src/memberships/types/casting.ts
@@ -1,0 +1,24 @@
+import { BlockFieldsFragment, MemberFieldsFragment } from '../queries'
+
+import { Block, Member } from './memberships'
+
+export const asBlock = (block: BlockFieldsFragment): Block => {
+  return {
+    ...block,
+  }
+}
+
+export const asMember = (data: MemberFieldsFragment): Member => {
+  return {
+    id: data.id,
+    handle: data.handle,
+    name: data.name ?? undefined,
+    avatar: data.avatarUri ?? undefined,
+    inviteCount: data.inviteCount,
+    isFoundingMember: data.isFoundingMember,
+    isVerified: data.isVerified,
+    rootAccount: data.rootAccount,
+    controllerAccount: data.controllerAccount,
+    roles: [],
+  }
+}
